Ordem de precedência

+++ PythonAulas/aula07a.py
@@ -8,8 +8,8 @@
 # n = input('Digite um nome: ')
 # print('Prazer em lhe conhecer, {}!'.format(n))
 
-n1 = int(input('Digite um valor: '))
-n2 = int(input('Digite um outro valor: '))
+n1 = float(input('Digite um valor: '))
+n2 = float(input('Digite um outro valor: '))
 s = n1 + n2
 m = n1 * n2
 d = n1 / n2
